Added other user's avatar to their most recently viewed message
Added the other user's avatar to SenderBubble.js, so the sender knows how far they have read.
Added a new property lastReadMessage to conversations, which is to track the most recently read message from the other user.
Updated updateReadMessage to readMessage in thunkCreators.js as its function name was the same as another function in reducerFunctions.js. This should remove confusion between the two.
Removed conversationId as a parameter from the series of functions related to readMessages as it isn't necessary since it can be retrieved from convoFragment. Also, the parameter conversation was renamed to convoFragment since its original name caused confusion since I assumed it had the same properties as convo used in reducerFunctions,js and what conversation normally has when passed as a parameter.
[KNOWN ISSUE] Currently, when a message is sent it does not update the lastReadMessage id to the most recent from the other user and persists until the other user sends a message, ie. the other user is always seen as one dialog exchange behind. This does not apply when the user selects the conversation to be active, and works as intended then.