Update Messages.js
Removed sorting from Messages.js. It appears to be sorted by the time it gets to the user.
Note: Using a fresh seed will have the messages appear out of order, but it will get sorted after selecting the conversation as active again, and this will reorder it for each newly selected existing conversation.