Added more security to reading messages
I changed messages/read route to use PUT instead of POST as well as changing the front-end request.
I removed numUpdates from messages.js as the returned number of updates from updates messages wasn't used. It could be used to minimize passing arrays at every request and only when an update is made, but it isn't currently.
I made reading messages more secure as the people reading the conversation must be part of the conversation to make the update.