Create conversation_user.js
Created a model to create a many-to-many associate between user and conversation, so conversations can have many users and users can have many conversations.
In addition to having the association, each user part of a conversation will have an unreadCount and lastReadMessage for that conversation to track where a user has read up to and if there are new messages sent.